Create debug.lua
